refactor Huffman example into library
